moved available golfers api endpoint to the seasons api
